Add files via upload
LRU (Least Recently Used) Cache is a cache which is implemented using two functions put(key, value) and get(key) where in the put operation, the key & value pair will get inserted into the cache until the cache's predetermined size isn't exceeding, and that too the key&value pair that is accessed now will become the most recently used pair and it will be pushed to the end of the cache. Where as in the get(key) function, returns -1 if the key that we are searching for isn't available and if key is available then it's corresponding value will be returned and that key&value pair now becomes most recently used pair and thus pushed to the end of cache.
Note: If we are trying to push a key&value pair if the cache is already full, then the least recently used pair will be removed(key&value pair present at the beginning of the cache).